resolved some PR comments
