+++ apps/web/.gitignore
@@ -41,3 +41,4 @@ yarn-error.log*
 next-env.d.ts
 
 package-lock.json
+pnpm-lock.yaml
