+++ apps/web/src/utils/news.ts
@@ -1,7 +1,17 @@
 import { Article } from "@/types/news.types";
-import { sanityQuery } from "../app/(public)/news/sanity";
 import { endpoints } from "@/data/endpoints";
-import { matchesTimeRange } from "@/utils/date";
+import { getDaysAgo } from "@/utils/date";
+
+export const matchesTimeRange = (articleDate: number, range: string) => {
+  if (range === "all") return true;
+  const age = getDaysAgo(articleDate);
+  if (range === "last-7") return age <= 7;
+  if (range === "last-30") return age <= 30;
+  if (range === "last-90") return age <= 90;
+  if (range === "this-year")
+    return new Date(articleDate).getFullYear() === new Date().getFullYear();
+  return true;
+};
 
 export const getFilteredArticles = (
   browseableArticles: Article[],
@@ -18,48 +28,8 @@ export const getFilteredArticles = (
     })
     .sort((l, r) => new Date(r.date).getTime() - new Date(l.date).getTime());
 
-export const escapeHtml = (value: string) =>
-  String(value ?? "")
-    .replaceAll("&", "&amp;")
-    .replaceAll("<", "&lt;")
-    .replaceAll(">", "&gt;")
-    .replaceAll('"', "&quot;")
-    .replaceAll("'", "&#39;");
-
 export const buildArticleUrl = (slug: string, articleDetailPageUrl: string) =>
   `${articleDetailPageUrl}?slug=${encodeURIComponent(String(slug).trim())}`;
-
-export async function fetchArticlesFromSantry(): Promise<Article[]> {
-  let articles: Article[] = [];
-
-  try {
-    const query = `
-      *[_type == "article" && status == "published"]
-      | order(publicationDate desc) {
-        "id": _id,
-        title,
-        "slug": slug.current,
-        author,
-        source,
-        "summary": excerpt,
-        field,
-        region,
-        readTime,
-        "imageUrl": coverImage.asset->url,
-        featured,
-        "date": publicationDate,
-        externalUrl,
-        sourceType
-      }
-    `;
-    articles = await sanityQuery(query);
-  } catch (err) {
-    console.error("Failed to load articles:", err);
-    throw new Error("Failed to load articles");
-  }
-
-  return articles;
-}
 
 export const getCategoryLabel = (
   category:
